Moved data models to subdir and updated imports and usings
