+++ src/BlazorGPT.Web/Pages/_ViewImports.cshtml
@@ -1,5 +1,3 @@
 ﻿@using Microsoft.AspNetCore.Identity
 
-
-GPT_Site.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
