+++ src/BlazorGPT.Web/_Imports.razor
@@ -12,6 +12,7 @@
 
 @using BlazorGPT
 @using BlazorGPT.Data
+@using BlazorGPT.Data.Model
 @using BlazorGPT.Pages
 @using BlazorGPT.Shared
 @using Radzen
