+++ src/BlazorGPT/Data/BlazorGptDBContext.cs
@@ -1,7 +1,5 @@
-﻿using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
+﻿using BlazorGPT.Data.Model;
 using Microsoft.EntityFrameworkCore;
-using OpenAI.GPT3.ObjectModels.RequestModels;
 
 namespace BlazorGPT.Data;
 
@@ -74,157 +72,5 @@ public class BlazorGptDBContext : DbContext
                     j.HasKey(cp => new { cp.ConversationId, cp.QuickProfileId });
                     j.ToTable("ConversationQuickProfiles");
                 });
-    }
-}
-
-public abstract class StateDataBase
-{
-    public Guid? Id { get; set; }
-    public string Name { get; set; } = null!;
-    public string? Content { get; set; }
-
-    public string? Type { get; set; } = null!;
-    public bool IsPublished { get; set; }
-}
-
-public class MessageState : StateDataBase {
-    public Guid? MessageId { get; set; }
-    public ConversationMessage? Message { get; set; }
-}
-
-public class ConversationTreeState : StateDataBase
-{
-    public Guid? ConversationId { get; set; }
-    public Conversation? Conversation { get; set; }
-}
-
-public class HiveState : StateDataBase
-{
-    public Guid? ConversationId { get; set; }
-    public Conversation? Conversation { get; set; }
-}
-
-public class ConversationMessage : ChatMessage
-{
-
-
-    public ConversationMessage(string role, string content) : base(role, content)
-    {
-    }
-
-    public ConversationMessage(ChatMessage msg) : base(msg.Role, msg.Content)
-    {
-    }
-
-    public Guid? Id { get; set; }
-    public Conversation? Conversation { get; set; }
-    public Guid? ConversationId { get; set; }
-
-    public MessageState? State { get; set; }
-
-    public ICollection<Conversation> BranchedConversations { get; set; } = new List<Conversation>();
-
-    public int? PromptTokens { get; set; }
-    public int? CompletionTokens { get; set; }
-
-    public DateTime Date { get; set; } = DateTime.UtcNow;
-
-}
-
-public class ConversationQuickProfile
-{
-    public Guid ConversationId { get; set; }
-    public Conversation Conversation { get; set; }
-
-    public Guid QuickProfileId { get; set; }
-    public QuickProfile QuickProfile { get; set; }
-}
-
-
-
-public class QuickProfile
-{
-    public Guid Id { get; set; }
-
-    public string Name { get; set; }
-
-    public string Content { get; set; }
-
-    public InsertAt InsertAt { get; set; }
-
-    public string? UserId { get; set; }
-
-    public bool EnabledDefault { get; set; }
-    public List<Conversation>? Conversations { get; set; }
-}
-
-
-public class Conversation
-{
-    public Guid? Id { get; set; }
-
-    [Required]
-    public string Model { get; set; } = null!;
-
-    [Required]
-    public string? UserId { get; set; } = null!;
-
-    public List<ConversationMessage> Messages { get; set; } = new List<ConversationMessage>();
-
-    public List<QuickProfile> QuickProfiles { get; set; } = new List<QuickProfile>();
-
-
-    public string? Summary { get; set; }
-
-
-    public DateTime DateStarted { get; set; }
-
-    public Guid? BranchedFromMessageId { get; set; }
-    public ConversationMessage? BranchedFromMessage { get; set; }
-
-    public List<ConversationTreeState> TreeStateList { get; set; } = new();
-
-    [NotMapped]
-    public Conversation HiveConversation { get; set; }
-
-    public HiveState? HiveState
-    { get; set; }
-
-    public void AddMessage(ConversationMessage message)
-    {
-        message.ConversationId = Id;
-        Messages.Add(message);
-    }
-
-
-    public static Conversation CreateConversation(string userId, string systemMessage, string? message = null)
-    {
-        Conversation conversation = new Conversation
-        {
-            UserId = userId,
-            DateStarted = DateTime.Now
-        };
-      
-        conversation.AddMessage(new ConversationMessage("system", systemMessage));
-        if (message != null)
-        {
-            conversation.AddMessage(new ConversationMessage("user", message));
-        }
-        return conversation;
-    }
- 
-}
-
-public static class DataExtensions
-{
-    public static bool HasStarted(this Conversation conversation)
-    {
-        return conversation.Messages.Any(m => m.Role == "assistant");
-    }
-
-    public static bool InitStage(this Conversation conversation)
-    {
-        return conversation.Messages.Last().Role == "assistant"
-               && conversation.Messages.Count(m => m.Role == "assistant") == 1;
     }
 }
