--- src/BlazorGPT/Data/Conversation.cs
@@ -1,2 +1,0 @@
-﻿namespace BlazorGPT.Data;
-
