--- src/BlazorGPT/Data/ConversationMessage.cs
@@ -1,2 +1,0 @@
-﻿namespace BlazorGPT.Data;
-
