+++ src/BlazorGPT/Data/Model/ConversationTreeState.cs
@@ -1,0 +1,7 @@
+﻿namespace BlazorGPT.Data.Model;
+
+public class ConversationTreeState : StateDataBase
+{
+    public Guid? ConversationId { get; set; }
+    public Conversation? Conversation { get; set; }
+}
