+++ src/BlazorGPT/Data/Model/HiveState.cs
@@ -1,0 +1,7 @@
+﻿namespace BlazorGPT.Data.Model;
+
+public class HiveState : StateDataBase
+{
+    public Guid? ConversationId { get; set; }
+    public Conversation? Conversation { get; set; }
+}
