+++ src/BlazorGPT/Data/Model/MessageState.cs
@@ -1,0 +1,6 @@
+﻿namespace BlazorGPT.Data.Model;
+
+public class MessageState : StateDataBase {
+    public Guid? MessageId { get; set; }
+    public ConversationMessage? Message { get; set; }
+}
