+++ src/BlazorGPT/GlobalUsings.cs
@@ -1,2 +1,3 @@
 ﻿global using Microsoft.EntityFrameworkCore;
 global using BlazorGPT.Data;
+global using BlazorGPT.Data.Model;
