+++ src/BlazorGPT/Managers/InterceptorBase.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text.RegularExpressions;
 using BlazorGPT.Data;
+using BlazorGPT.Data.Model;
 
 namespace BlazorGPT.Managers;
