+++ src/BlazorGPT/Managers/InterceptorHandler.cs
@@ -1,4 +1,5 @@
 ﻿using BlazorGPT.Data;
+using BlazorGPT.Data.Model;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
